backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
gunicorn==23.0.0
psycopg2-binary==2.9.9
python-multipart==0.0.9
google-generativeai>=0.8.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx>=0.27.0
numpy>=1.26.0
openai>=1.0.0
